# cert-api2/public/swagger.yml
openapi: "3.0.0"
info:
  version: 0.0.1
  title: Taichung School Certificate Project
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api/v1
  - url: http://localhost:3000/api/test
  - url: http://192.168.56.101:3000/api/v1
security:
  - bearerAuth: []
paths:
...
  /certtest/issue/get-class-list:       
    post:     
      summary: get class list by schoolID
      tags:
        - issue
      parameters: []        
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                userID:
                  type: number
                orgID:
                  type: number
                classID:
                  type: number
                role:
                  type: number
                address:
                  type: string    
              example:
                userID: -1
                orgID: 55688
                classID: 10
                role: -1
                address: null 
      responses:
        '200':
          content:
            application/json:      
  /certtest/issue/get-student-list:         
    post:        
      summary: get student list by schoolID and class
      tags:
        - issue
      parameters:
        - in: query
          name: schoolID
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
              example:
                schoolID: "NCTU"
      responses:
        '200':
          content:
            application/json:       
  /certtest/issue:       
    post:     
      summary: issue main
      tags:
        - issue
      parameters: []        
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                owner:
                  type: string[]
                issuer:
                  type: string
                signer:
                  type: string[]
                contract:
                  type: string
                version:
                  type: string    
                certName:
                  type: string 
                revocation:
                  type: string 
                description:
                  type: string 
                timestamp:
                  type: string 
                certFile:
                  type: string[]
              example:
                owner: ["9417158","9417159","9417160"]
                issuer: "issuer123"
                signer: ["9857503","9857504","9857505"]
                contract: "0686005"
                version: "0.1.0"
                certName: "TOEFL"
                revocation: false
                description: "this is toefl certification"
                timestamp: "1234567890123"
                certFilePath: ["/home/superorange/9417158.pdf",
                    "/home/superorange/9417159.pdf",
                    "/home/superorange/9417160.pdf"]
      responses:
        '200':
          content:
            application/json:      
  /certtest/users:
    get:
      summary: get student list
      tags:
        - certtest
      parameters: []
      responses:
        '200':
          content:
            application/json:
    post:
      summary: get classID by schoolID
      tags:
        - certtest
      parameters:
        - in: query
          name: schoolID
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string[]
              example:
                schoolID: "NCTU"
      responses:
        '200':
          content:
            application/json:
...
  /dev/bob/blockreq:
    get:
      summary: dev, blocking request 
      tags:
        - dev
      parameters:
        - in: query
          name: seconds 
          schema:
            type: integer
          examples:
            seconds10:
              value: 10
      responses:
        '200':
          content:
            application/json:
  /dev/bob/address:
    post:
      summary: dev, bob's address
      tags:
        - dev
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type: string
                id: number
              example:
                id: 1009
                type: SYSTEM
      responses:
        '200':
          content:
            application/json:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...